Uncracked keys df: stop dropping records when a pair (r, pubkey) has been cracked
Dropping the records improve the performance, but lead to unoptimal attacks.

Example: we have 5 sigs for 2 pubkeys p1 and p2 that occured between 2016  and 2018
- (r1 s1) p1 2016
- (r1, s2) p1 2018
- (r1, s3) p2 2017
- (r2, s4) p2 2017
- (r2, s5) p1 2017

We can crack (r1, p1) because of the repeated nonce. If we drop the 2 first records, we won't see the
possibility to crack p2 in 2017. We'll have to suppose that r1 was known only in 2018 and hence that p2 could be cracked at best in 2018.
While because of the sys composed of the last 4 equations, p2 and p1 can be cracked as soon as 2017.